Firefly-866: response from feedback
