+++ buildScript/tasks.gincl
@@ -131,13 +131,12 @@ task createVersionTag  {
         props.setProperty('BuildTime', build_time)
         props.setProperty('BuildTag', tag)
         props.setProperty('BuildCommit', getCommitHash())
-        props.setProperty('BuildGitTagFirefly', getCommitTag())
+        props.setProperty('BuildGitTagFirefly', getCommitTag(fireflyPath))
 
         props.setProperty('ExtendedCache', project.env == "local" ? (project.hasProperty("ExtendedCache") ? project.ExtendedCache : "true") : "false")
 
         if (fireflyPath != rootDir.getPath()) {
             props.setProperty('BuildCommitFirefly', getCommitHash(fireflyPath))
-            props.setProperty('BuildGitTagFirefly', getCommitTag(fireflyPath))
         }
 
         props.store(file("${project.buildDir}/version.tag").newWriter(), "Version Info")
